requirement.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Werkzeug==1.0.1